docker optimization

+++ packaging/docker/bootstrap_scripts/mongodb.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-echo "# INSTALL MONGODB"
-export DEBIAN_FRONTEND=noninteractive
-apt-get update
-apt-get install -y mongodb
+if [ "$DEV" == "yes" ]
+then
+    echo "# INSTALL MONGODB"
+    export DEBIAN_FRONTEND=noninteractive
+    apt-get update
+    apt-get install --no-install-recommends -y mongodb
 
-# apt cleanup
-rm -rf /var/lib/apt/lists/*
-apt-get clean
+    # apt cleanup
+    rm -rf /var/lib/apt/lists/*
+    apt-get clean
+fi
